Progress: Revision CRUD Profile Kontak Page

+++ app/Http/Controllers/KontakController.php
@@ -78,8 +78,10 @@ class KontakController extends Controller
         ]);
 
         if ($request->file('icon')) {
-            $oldImagePath = public_path('assets/img/profil-images/kontak-image/') . $request->oldImage;
-            unlink($oldImagePath);
+            if (public_path('assets/img/profil-images/kontak-image/') . $request->oldImage && $request->oldImage) {
+                $oldImagePath = public_path('assets/img/profil-images/kontak-image/') . $request->oldImage;
+                unlink($oldImagePath);
+            }
 
             $image = $request->file('icon');
             $imageName = time() . '.' . $image->getClientOriginalExtension();
@@ -103,8 +105,10 @@ class KontakController extends Controller
         $contact = Contact::where('id', $id)->first();
 
         if ($contact->icon) {
-            $imagePath = public_path('assets/img/profil-images/kontak-image/') . $contact->icon;
-            unlink($imagePath);
+            if (public_path('assets/img/profil-images/kontak-image/') . $contact->icon && $contact->icon) {
+                $imagePath = public_path('assets/img/profil-images/kontak-image/') . $contact->icon;
+                unlink($imagePath);
+            }
         }
 
         $contact = $contact->delete();
